Make multi run available for ArtemisIn*
