+++ tlr/tests-tlr/src/test/java/edu/kit/kastel/mcse/ardoco/tlr/tests/integration/ArtemisIT.java
@@ -2,9 +2,7 @@
 package edu.kit.kastel.mcse.ardoco.tlr.tests.integration;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Locale;
 import java.util.stream.Stream;
 
 import org.eclipse.collections.api.factory.Lists;
@@ -17,8 +15,6 @@ import org.junit.jupiter.api.condition.EnabledIfEnvironmentVariable;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import edu.kit.kastel.mcse.ardoco.core.api.output.ArDoCoResult;
 import edu.kit.kastel.mcse.ardoco.core.common.util.Environment;
@@ -30,7 +26,6 @@ import edu.kit.kastel.mcse.ardoco.tlr.tests.integration.evaluation.ArtemisEvalua
 
 class ArtemisIT extends AbstractArdocoIT {
     private static final int NUMBER_OF_RUNS = 5;
-    protected Logger logger = LoggerFactory.getLogger(this.getClass());
 
     @BeforeAll
     static void beforeAll() {
@@ -72,54 +67,6 @@ class ArtemisIT extends AbstractArdocoIT {
             results.add(result);
         }
         averageAndLog(results);
-    }
-
-    private void averageAndLog(List<SingleClassificationResult<String>> results) {
-        var precisions = results.stream().mapToDouble(SingleClassificationResult::getPrecision).toArray();
-        var recalls = results.stream().mapToDouble(SingleClassificationResult::getRecall).toArray();
-        var f1s = results.stream().mapToDouble(SingleClassificationResult::getF1).toArray();
-        var accuracies = results.stream().mapToDouble(SingleClassificationResult::getAccuracy).toArray();
-        var specificities = results.stream().mapToDouble(SingleClassificationResult::getSpecificity).toArray();
-        var phis = results.stream().mapToDouble(SingleClassificationResult::getPhiCoefficient).toArray();
-
-        var avgPrecision = Arrays.stream(precisions).average().orElse(0.0);
-        var avgRecall = Arrays.stream(recalls).average().orElse(0.0);
-        var avgF1 = Arrays.stream(f1s).average().orElse(0.0);
-        var avgAccuracy = Arrays.stream(accuracies).average().orElse(0.0);
-        var avgSpecificity = Arrays.stream(specificities).average().orElse(0.0);
-        var avgPhi = Arrays.stream(phis).average().orElse(0.0);
-
-        var stdPrecision = stddev(precisions, avgPrecision);
-        var stdRecall = stddev(recalls, avgRecall);
-        var stdF1 = stddev(f1s, avgF1);
-        var stdAccuracy = stddev(accuracies, avgAccuracy);
-        var stdSpecificity = stddev(specificities, avgSpecificity);
-        var stdPhi = stddev(phis, avgPhi);
-
-        if (logger.isInfoEnabled()) {
-            var logString = String.format(Locale.ENGLISH, """
-
-                    Average results (± StdDev):
-                    	Precision:   %8.2f ± %6.2f
-                    	Recall:      %8.2f ± %6.2f
-                    	F1:          %8.2f ± %6.2f
-                    	Accuracy:    %8.2f ± %6.2f
-                    	Specificity: %8.2f ± %6.2f
-                    	Phi Coef.:   %8.2f ± %6.2f
-                    """, avgPrecision, stdPrecision, avgRecall, stdRecall, avgF1, stdF1, avgAccuracy, stdAccuracy, avgSpecificity, stdSpecificity, avgPhi,
-                    stdPhi);
-            logger.info(logString);
-        }
-    }
-
-    private static double stddev(double[] values, double mean) {
-        if (values.length == 0)
-            return 0.0;
-        double sum = 0.0;
-        for (double v : values) {
-            sum += (v - mean) * (v - mean);
-        }
-        return Math.sqrt(sum / values.length);
     }
 
     public SingleClassificationResult<String> runTraceLinkEvaluation(ArtemisEvaluationProject project, LargeLanguageModel llm) {
