+++ tlr/tests-tlr/src/test/java/edu/kit/kastel/mcse/ardoco/tlr/tests/integration/ArtemisInExArchIT.java
@@ -3,43 +3,35 @@ package edu.kit.kastel.mcse.ardoco.tlr.tests.integration;
 
 import static edu.kit.kastel.mcse.ardoco.core.api.models.Metamodel.CODE_WITH_COMPILATION_UNITS;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Locale;
-import java.util.Map;
-import java.util.stream.Collectors;
 
-import org.eclipse.collections.api.tuple.Pair;
-import org.eclipse.collections.impl.tuple.Tuples;
+import org.eclipse.collections.api.factory.Lists;
 import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Assumptions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.condition.DisabledIfEnvironmentVariable;
+import org.junit.jupiter.api.condition.EnabledIfEnvironmentVariable;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import edu.kit.kastel.mcse.ardoco.core.api.output.ArDoCoResult;
 import edu.kit.kastel.mcse.ardoco.core.common.util.Environment;
-import edu.kit.kastel.mcse.ardoco.metrics.ClassificationMetricsCalculator;
-import edu.kit.kastel.mcse.ardoco.metrics.result.AggregationType;
 import edu.kit.kastel.mcse.ardoco.metrics.result.SingleClassificationResult;
 import edu.kit.kastel.mcse.ardoco.tlr.models.informants.LargeLanguageModel;
 import edu.kit.kastel.mcse.ardoco.tlr.models.informants.LlmArchitecturePrompt;
 import edu.kit.kastel.mcse.ardoco.tlr.tests.approach.ArDoCodeEvaluationProject;
 import edu.kit.kastel.mcse.ardoco.tlr.tests.integration.evaluation.ArtemisInExArchEvaluation;
-import edu.kit.kastel.mcse.ardoco.tlr.tests.integration.evaluation.ExArchEvaluation;
 
 @Disabled("Only for manual execution")
-class ArtemisInExArchIT {
+class ArtemisInExArchIT extends AbstractArdocoIT {
+    private static final int NUMBER_OF_RUNS = 5;
+
     private static final Logger logger = LoggerFactory.getLogger(ArtemisInExArchIT.class);
     protected static final String LOGGING_ARDOCO_CORE = "org.slf4j.simpleLogger.log.edu.kit.kastel.mcse.ardoco.core";
-
-    private static final Map<Pair<ArDoCodeEvaluationProject, LargeLanguageModel>, ArDoCoResult> RESULTS = new HashMap<>();
 
     @BeforeAll
     static void beforeAll() {
@@ -52,6 +44,7 @@ class ArtemisInExArchIT {
         System.setProperty(LOGGING_ARDOCO_CORE, "error");
     }
 
+    @DisabledIfEnvironmentVariable(named = "mutipleRuns", matches = ".*")
     @DisplayName("Evaluate Artemis @ ExArch (SAD-SAM-via-LLM-Code) TLR")
     @ParameterizedTest(name = "{0}")
     @EnumSource(ArDoCodeEvaluationProject.class)
@@ -67,57 +60,47 @@ class ArtemisInExArchIT {
 
         LargeLanguageModel llmForNer = LargeLanguageModel.GPT_5;
 
-        logger.info("###############################################");
         logger.info("Evaluating project {} with LLM '{}', '{}'", project, llmForExArch, llmForNer);
         logger.info("Prompts: {}, {}, {}", docPrompt, codePrompt, aggPrompt);
         logger.info("Features: {}", codeFeatures);
 
         var evaluation = new ArtemisInExArchEvaluation(project, llmForExArch, docPrompt, codePrompt, codeFeatures, aggPrompt, llmForNer);
         var result = evaluation.runTraceLinkEvaluation();
-        if (result != null) {
-            RESULTS.put(Tuples.pair(project, llmForExArch), result);
-        }
-        System.err.printf("--- Evaluated project %s with LLM '%s' ---%n", project, llmForExArch);
-        logger.info("###############################################");
+        Assertions.assertNotNull(result);
     }
 
-    @AfterAll
-    static void printResults() {
-        System.setProperty("org.slf4j.simpleLogger.log.edu.kit.kastel.mcse.ardoco.tlr.tests.integration.evaluation", "error");
-        System.out.println(Arrays.stream(ArDoCodeEvaluationProject.values())
-                .map(Enum::name)
-                .collect(Collectors.joining(" & ")) + " & Macro Avg & Weighted Average" + " \\\\");
-        for (LargeLanguageModel llm : List.of(LargeLanguageModel.GPT_4_O)) {
-            if (llm.isGeneric() && RESULTS.keySet().stream().noneMatch(k -> k.getTwo() == llm)) {
-                continue;
-            }
-            StringBuilder llmResult = new StringBuilder(llm.getHumanReadableName() + " ");
+    @EnabledIfEnvironmentVariable(named = "mutipleRuns", matches = ".*")
+    @DisplayName("Evaluate Artemis @ ExArch (SAD-SAM-via-LLM-Code) TLR Multi")
+    @ParameterizedTest(name = "{0}")
+    @EnumSource(ArDoCodeEvaluationProject.class)
+    void evaluateArtemisInExArchMultiple(ArDoCodeEvaluationProject project) {
+        Assumptions.assumeTrue(Environment.getEnv("CI") == null);
 
-            List<SingleClassificationResult<String>> classificationResults = new ArrayList<>();
-            for (ArDoCodeEvaluationProject project : ArDoCodeEvaluationProject.values()) {
-                if (!RESULTS.containsKey(Tuples.pair(project, llm))) {
-                    llmResult.append("&--&--&--");
-                    continue;
-                }
-                ArDoCoResult result = RESULTS.get(Tuples.pair(project, llm));
+        LargeLanguageModel llmForExArch = LargeLanguageModel.GPT_4_O;
+        LlmArchitecturePrompt docPrompt = LlmArchitecturePrompt.EXTRACT_FROM_ARCHITECTURE;
+        LlmArchitecturePrompt codePrompt = null;
+        LlmArchitecturePrompt aggPrompt = null;
 
-                var goldStandard = project.getTlrTask().getExpectedTraceLinks();
-                goldStandard = ExArchEvaluation.enrollGoldStandard(goldStandard, result, CODE_WITH_COMPILATION_UNITS);
-                var evaluationResults = ExArchEvaluation.calculateEvaluationResults(result, goldStandard, CODE_WITH_COMPILATION_UNITS);
-                classificationResults.add(evaluationResults);
-                llmResult.append(String.format(Locale.ENGLISH, "&%.2f&%.2f&%.2f", evaluationResults.getPrecision(), evaluationResults.getRecall(),
-                        evaluationResults.getF1()));
-            }
-            ClassificationMetricsCalculator classificationMetricsCalculator = ClassificationMetricsCalculator.getInstance();
-            var averages = classificationMetricsCalculator.calculateAverages(classificationResults, null);
+        LlmArchitecturePrompt.Features codeFeatures = LlmArchitecturePrompt.Features.PACKAGES;
 
-            var macro = averages.stream().filter(it -> it.getType() == AggregationType.MACRO_AVERAGE).findFirst().orElseThrow();
-            var weighted = averages.stream().filter(it -> it.getType() == AggregationType.WEIGHTED_AVERAGE).findFirst().orElseThrow();
+        LargeLanguageModel llmForNer = LargeLanguageModel.GPT_5;
 
-            llmResult.append(String.format(Locale.ENGLISH, "&%.2f&%.2f&%.2f&%.2f&%.2f&%.2f\\\\", macro.getPrecision(), macro.getRecall(), macro.getF1(),
-                    weighted.getPrecision(), weighted.getRecall(), weighted.getF1())); // end of line
-            System.out.println(llmResult.toString().replace("0.", ".").replace("1.00", "1.0"));
-            System.setProperty("org.slf4j.simpleLogger.log.edu.kit.kastel.mcse.ardoco.tlr.tests.integration.evaluation", "info");
+        logger.info("Evaluating project {} with LLM '{}', '{}'", project, llmForExArch, llmForNer);
+        logger.info("Prompts: {}, {}, {}", docPrompt, codePrompt, aggPrompt);
+        logger.info("Features: {}", codeFeatures);
+
+        List<SingleClassificationResult<String>> results = Lists.mutable.empty();
+        for (int i = 0; i < NUMBER_OF_RUNS; i++) {
+            logger.info("Eval run {}/{} [{}]", i, NUMBER_OF_RUNS, project);
+            var evaluation = new ArtemisInExArchEvaluation(project, llmForExArch, docPrompt, codePrompt, codeFeatures, aggPrompt, llmForNer);
+            var result = evaluation.runTraceLinkEvaluation();
+            Assertions.assertNotNull(result);
+
+            var goldStandard = project.getTlrTask().getExpectedTraceLinks();
+            goldStandard = ArtemisInExArchEvaluation.enrollGoldStandard(goldStandard, result, CODE_WITH_COMPILATION_UNITS);
+            var evalResult = ArtemisInExArchEvaluation.calculateEvaluationResults(result, goldStandard, CODE_WITH_COMPILATION_UNITS);
+            results.add(evalResult);
         }
+        averageAndLog(results);
     }
 }
